# api/models.py
# generated by fastapi-codegen:
#   filename:  swagger\object_api.yaml
#   timestamp: 2021-04-09T15:12:24+00:00

from __future__ import annotations

from datetime import datetime
from typing import Any, Dict, Optional
from uuid import UUID

from pydantic import BaseModel, Field


class Object(BaseModel):
    id: UUID = Field(..., example='d290f1ee-6c54-4b01-90e6-d701748f0851')
    path: Optional[str] = Field(
        None, example='/data/d290f1ee-6c54-4b01-90e6-d701748f0851.mp4'
    )
    mimetype: Optional[str] = Field(None, example='video/mp4')
    size: Optional[int] = Field(None, example=234233342)
    date: Optional[datetime] = Field(None, example='2016-08-29T09:12:33.001Z')
    creation: Optional[datetime] = Field(None, example='2016-08-29T09:12:33.001Z')
    meta: Optional[Dict[str, Any]] = Field(
        None,
        example={'size': {'width': 640, 'height': 480}, 'fps': 10, 'device': 'radarpi'},
    )